requirements.txt
slack_sdk == 3.13.0
python-dotenv == 0.19.0
requests == 2.22.0